+++ openquake/hazardlib/geo/mesh.py
@@ -98,27 +98,6 @@ def sqrt(array):
     return numpy.sqrt(array)
 
 
-def surface_to_arrays(surface):
-    """
-    :param surface: a (Multi)Surface object
-    :returns: a list of S arrays of shape (3, N, M)
-    """
-    if hasattr(surface, 'surfaces'):  # multiplanar surfaces
-        lst = []
-        for surf in surface.surfaces:
-            arr = surf.mesh.array
-            if len(arr.shape) == 2:  # PlanarSurface
-                arr = arr.reshape(3, 1, 4)
-            lst.append(arr)
-        return lst
-    mesh = surface.mesh
-    if len(mesh.lons.shape) == 1:  # 1D mesh
-        shp = (3, 1) + mesh.lons.shape
-    else:  # 2D mesh
-        shp = (3,) + mesh.lons.shape
-    return [mesh.array.reshape(shp)]
-
-
 def calc_inclination(earth_surface_tangent_normal, tl_normal, tl_area,
                      br_normal, br_area):
     # inclination calculation
