+++ openquake/hazardlib/source/rupture.py
@@ -29,8 +29,8 @@ import json
 from openquake.baselib import general, hdf5
 from openquake.hazardlib import geo, site, scalerel
 from openquake.hazardlib.geo.nodalplane import NodalPlane
-from openquake.hazardlib.geo.mesh import (
-    Mesh, RectangularMesh, surface_to_arrays)
+from openquake.hazardlib.geo.mesh import Mesh, RectangularMesh
+from openquake.hazardlib.geo.surface.base import surface_to_arrays, to_arrays
 from openquake.hazardlib.geo.point import Point
 from openquake.hazardlib.geo.geodetic import geodetic_distance
 from openquake.hazardlib.near_fault import (
@@ -129,23 +129,6 @@ def to_csv_array(ebruptures):
         _fixfloat32(extra)
         rec['extra'] = json.dumps(extra)
     return arr
-
-
-def to_arrays(geom):
-    """
-    :param geom: an array [num_surfaces, shape_y, shape_z ..., coords]
-    :returns: a list of num_surfaces arrays with shape (3, shape_y, shape_z)
-    """
-    arrays = []
-    num_surfaces = int(geom[0])
-    start = num_surfaces * 2 + 1
-    for i in range(1, 2 * num_surfaces, 2):
-        s1, s2 = int(geom[i]), int(geom[i + 1])
-        size = s1 * s2 * 3
-        array = geom[start:start + size].reshape(3, s1, s2)
-        arrays.append(array)
-        start += size
-    return arrays
 
 
 def get_ebr(rec, geom, trt):
